+++ app/src/main/java/com/eps/todoturtle/mock/MockValues.kt
@@ -2,7 +2,8 @@ package com.eps.todoturtle.mock
 
 enum class MockValues(private val value: String) {
     USERNAME("user"),
-    PASSWORD("password");
+    PASSWORD("password"),
+    ;
 
     fun getValue(): String {
         return value
