+++ app/src/main/java/com/eps/todoturtle/permissions/PermissionDialog.kt
@@ -21,7 +21,6 @@ fun PermissionDialog(
     onDismiss: () -> Unit,
     onOkClick: () -> Unit,
     onGoToAppSettingsClick: () -> Unit,
-    modifier: Modifier = Modifier,
 ) {
     AlertDialog(
         onDismissRequest = onDismiss,
@@ -49,6 +48,6 @@ fun PermissionDialog(
                 text = permissionTextProvider.getDescription(isPermanentlyDeclined),
             )
         },
-        modifier = modifier,
+        modifier = Modifier,
     )
 }
