+++ app/src/main/java/com/eps/todoturtle/permissions/PermissionsViewModel.kt
@@ -3,7 +3,7 @@ package com.eps.todoturtle.permissions
 import androidx.compose.runtime.mutableStateListOf
 import androidx.lifecycle.ViewModel
 
-class PermissionsViewModel: ViewModel() {
+class PermissionsViewModel : ViewModel() {
     val declinedPermissionQueue = mutableStateListOf<String>()
 
     fun onDismissDialog() {
