+++ app/src/main/java/com/eps/todoturtle/permissions/providers/CameraPermissionProvider.kt
@@ -3,7 +3,7 @@ package com.eps.todoturtle.permissions.providers
 import android.Manifest
 import com.eps.todoturtle.permissions.providers.text.CameraTextProvider
 
-class CameraPermissionProvider: PermissionProvider {
+class CameraPermissionProvider : PermissionProvider {
     override val textProvider = CameraTextProvider()
     override val permission = Manifest.permission.CAMERA
 }
