+++ app/src/main/java/com/eps/todoturtle/ui/MainActivityUI.kt
@@ -3,7 +3,6 @@ package com.eps.todoturtle.ui
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
