+++ app/src/main/java/com/eps/todoturtle/ui/profile/details/SignOutProfileButton.kt
@@ -1,12 +1,10 @@
 package com.eps.todoturtle.ui.profile.details
 
-import android.widget.Toast
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
