+++ app/src/main/java/com/eps/todoturtle/ui/profile/shared/OutlinedText.kt
@@ -1,6 +1,5 @@
 package com.eps.todoturtle.ui.profile.shared
 
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
